remove go-http-auth library
It requires password storage, which would not work well with other backends
(eg. LDAP). I tried to refactor it but because of the support for Digest auth
it becomes messy.
So take some of the good parts (htpasswd) and add our own http auth mini-library.
This one is based on a custom check function.